+++ media-wrangler-front-end/package.json
@@ -16,16 +16,14 @@
     "@mui/lab": "^6.0.0-beta.19",
     "@mui/material": "^6.2.0",
     "axios": "^1.7.9",
-    "bulma": "^1.0.2",
-    "formik": "^2.4.6",
     "bulma": "^1.0.3",
+    "formik": "^2.4.6",
     "mui": "^0.0.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^7.1.1",
     "react-icons": "^5.4.0",
+    "react-router-dom": "^7.1.1",
     "react-star-ratings": "^2.3.0"
-
   },
   "devDependencies": {
     "@eslint/js": "^9.15.0",
